refactor: offset 값 1 -> 0 으로 수정
